backend/requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
websockets==12.0

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.31
alembic==1.13.2

# AI Providers
anthropic==0.28.0
openai==1.35.0
google-genai==0.8.0
groq==0.9.0
cohere==5.5.8
mistralai==1.0.0

# HTTP / Async
httpx==0.27.0
aiohttp==3.9.5

# DNS / Network
dnspython==2.6.1
python-whois==0.9.4

# Data Validation
pydantic==2.7.1
pydantic-settings==2.3.0

# Utilities
python-dotenv==1.0.1
loguru==0.7.2

# Security
cryptography==42.0.8

# Optional OSINT
shodan==1.31.0
